Refactor sidebar sections/paths code
