+++ src/lib/paths/paths.ts
