+++ src/lib/recoil/state-effects/use-state-effect.ts
@@ -1,17 +1,17 @@
-import { useEffect } from 'react';
-import { RecoilState, useRecoilCallback, useRecoilValue } from 'recoil';
+import { useEffect, useLayoutEffect } from 'react';
+import {
+  RecoilState,
+  RecoilValue,
+  RecoilValueReadOnly,
+  useRecoilCallback,
+  useRecoilValue,
+} from 'recoil';
 
 import { usePrevious } from '@/lib/hooks/use-previous';
 
 import { StateEffect } from './types';
 
-/**
- * Run a state effect when a piece of state changes.
- * A state effect can modify other pieces of state.
- * @param state the recoil state to watch
- * @param effect the state effect to run when the state changes
- */
-export function useStateEffect<T>(state: RecoilState<T>, effect: StateEffect<T>) {
+function useStateEffectImpl<T>(state: RecoilValue<T>, effect: StateEffect<T>) {
   const stateValue = useRecoilValue(state);
 
   const previousStateValue = usePrevious(stateValue);
@@ -24,5 +24,29 @@ export function useStateEffect<T>(state: RecoilState<T>, effect: StateEffect<T>)
     [effect],
   );
 
+  return [stateValue, previousStateValue, cb] as const;
+}
+
+/**
+ * Run a state effect when a piece of state changes.
+ * A state effect can modify other pieces of state.
+ * @param state the recoil state to watch
+ * @param effect the state effect to run when the state changes
+ */
+export function useStateEffect<T>(state: RecoilValueReadOnly<T>, effect: StateEffect<T>) {
+  const [stateValue, previousStateValue, cb] = useStateEffectImpl(state, effect);
+
   useEffect(() => cb(stateValue, previousStateValue), [cb, stateValue, previousStateValue]);
+}
+
+/**
+ * Run a state layout effect when a piece of state changes.
+ * A state effect can modify other pieces of state.
+ * @param state the recoil state to watch
+ * @param effect the state effect to run when the state changes
+ */
+export function useStateLayoutEffect<T>(state: RecoilValueReadOnly<T>, effect: StateEffect<T>) {
+  const [stateValue, previousStateValue, cb] = useStateEffectImpl(state, effect);
+
+  useLayoutEffect(() => cb(stateValue, previousStateValue), [cb, stateValue, previousStateValue]);
 }
